Revert "Trying to figure out z-buffering issue"
This reverts commit 22edae8294a7f05e31b17c94046d230abecedb14.

+++ GaleEngine/Rendering/Texture.cpp
@@ -19,7 +19,7 @@ void Texture::Destroy() {
 
 Texture::Texture(const string &filename) {
 	name = filename;
-	unsigned int dotPos = (unsigned int)filename.find('.');
+	unsigned int dotPos = filename.find('.');
 	string extension = filename.substr(dotPos);
 
 	if (extension == ".bmp" || extension == ".BMP") {
